refact: use Robolectric at :platform:testing
