+++ platform/autos/build.gradle.kts
@@ -39,7 +39,6 @@ dependencies {
   testImplementation(project(":platform:autos-test"))
   testImplementation(project(":platform:testing"))
   testImplementation(libs.android.compose.ui.test.manifest)
-  testImplementation(libs.android.test.core)
   testImplementation(libs.assertk)
   testImplementation(libs.kotlin.coroutines.test)
   testImplementation(libs.kotlin.test)
