+++ platform/testing/src/test/AndroidManifest.xml
