+++ platform/testing/src/test/java/br/com/orcinus/orca/platform/testing/InstrumentationExtensionsTests.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2024 Orcinus
+ * Copyright © 2023–2024 Orcinus
  *
  * This program is free software: you can redistribute it and/or modify it under the terms of the
  * GNU General Public License as published by the Free Software Foundation, either version 3 of the
@@ -13,22 +13,19 @@
  * not, see https://www.gnu.org/licenses.
  */
 
-package br.com.orcinus.orca.platform.testing.activity.scenario
+package br.com.orcinus.orca.platform.testing
 
-import android.app.Activity
-import androidx.test.ext.junit.rules.ActivityScenarioRule
+import androidx.test.platform.app.InstrumentationRegistry
 import assertk.assertThat
 import assertk.assertions.isSameAs
-import org.junit.Rule
 import org.junit.Test
+import org.junit.runner.RunWith
+import org.robolectric.RobolectricTestRunner
 
-internal class ActivityScenarioExtensionsTests {
-  @get:Rule val activityScenarioRule = ActivityScenarioRule(TestActivity::class.java)
-
-  class TestActivity : Activity()
-
+@RunWith(RobolectricTestRunner::class)
+internal class InstrumentationExtensionsTests {
   @Test
-  fun getsActivity() {
-    activityScenarioRule.scenario?.run { onActivity { assertThat(activity).isSameAs(it) } }
+  fun contextIsThatOfInstrumentationFromRegistry() {
+    assertThat(context).isSameAs(InstrumentationRegistry.getInstrumentation().context)
   }
 }
