+++ platform/testing/src/test/res/values/strings.xml
